je monge la poisson

+++ math/0x03-probability/poisson.py
@@ -32,4 +32,4 @@ class Poisson:
                 TypeError('data must be a list')
             if len(data) < 2:
                 ValueError('data must contain multiple values')
-            self.lambtha = mean(data)
+            self.lambtha = float(mean(data))
